requirements.txt
pytests==8.4.2